refactor(otlp-transformer): de-dupe resource and instrumentation scope code
